requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
agno>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
tavily-python
elevenlabs>=0.2.0
requests>=2.25.0
Pillow>=8.0.0
